+++ slash/loader.py
@@ -25,6 +25,8 @@ from .core.runnable_test_factory import RunnableTestFactory
 from .utils.pattern_matching import Matcher
 from .utils.python import check_duplicate_functions
 from .resuming import ResumedTestData
+from .utils.interactive import generate_interactive_test
+
 
 _logger = Logger(__name__)
 
@@ -51,7 +53,7 @@ class Loader(object):
         return self._cached_matchers
 
 
-    def get_runnables(self, paths):
+    def get_runnables(self, paths, prepend_interactive=False):
         assert context.session is not None
         sources = (t for repetition in range(config.root.run.repeat_all)
                    for t in self._generate_test_sources(paths))
@@ -59,6 +61,10 @@ class Loader(object):
         self._duplicate_funcs |= self._local_config.duplicate_funcs
         for (path, name, line) in sorted(self._duplicate_funcs):
             _logger.warning('Duplicate function definition, File: {}, Name: {}, Line: {}'.format(path, name, line))
+
+        if prepend_interactive:
+            returned.insert(0, generate_interactive_test())
+
         hooks.tests_loaded(tests=returned) # pylint: disable=no-member
         returned.sort(key=lambda test: test.__slash__.get_sort_key())
         return returned