+++ tests/test_fixture_exclusion.py
@@ -23,3 +23,20 @@ def test_fixture_exclusion():
     assert results[0].data['value'] == 10
     assert results[1].is_skip()
     assert not results[1].is_started()
+
+
+def test_fixture_exclusion_nested(suite, suite_test):
+
+    fixture = suite.slashconf.add_fixture()
+    param = fixture.add_parameter()
+    suite_test.depend_on_fixture(fixture)
+
+    param_value = param.values[0]
+
+    suite_test.add_decorator('slash.exclude("{fixture_name}.{param_name}", {values})'.format(
+        fixture_name=fixture.name,
+        param_name=param.name,
+        values=[param_value],
+    ))
+    suite_test.exclude_param_value(param.name, param_value)
+    suite.run()