+++ doc/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 * :feature:`-` Assertions coming from plugins and modules loaded from the project's ``.slashrc`` now also have assertion rewriting introspection enabled
 * :bug:`- major` Honor run.default_sources configuration when using slash list (thanks Pierre-Luc Tessier Gagné)
 * :bug:`- major` Several Windows-specific fixes (thanks Pierre-Luc Tessier Gagné)
+* :release:`1.2.4 <19-06-2017>`
+* :bug:`581` Fix ``slash.exclude`` to work across fixture namespaces
+* :bug:`580` ``tests_loaded`` hooks now get called with a list of tests including the interactive test if applicable
 * :release:`1.2.2 <29-05-2017>`
 * :bug:`564` Fix test collection bug causing tests to not be loaded with some plugins
 * :release:`1.2.0 <30-04-2017>`